# perf: Skip flattened optimization if we unpack tags anyway