--- a/snuba/query/processors/tagsmap.py
+++ b/snuba/query/processors/tagsmap.py
@@ -6,6 +6,7 @@ from enum import Enum
 from typing import Optional, List, NamedTuple, Set
 
 from snuba.datasets.tags_column_processor import NESTED_COL_EXPR_RE
+from snuba.query.expressions import Column, Expression
 from snuba.query.query import Query
 from snuba.query.query_processor import QueryProcessor
 from snuba.datasets.transactions_processor import escape_field
@@ -103,6 +104,14 @@ class NestedFieldConditionOptimizer(QueryProcessor):
             )
         return None
 
+    def __find_tags(self, expression: Expression) -> bool:
+        for node in expression:
+            if isinstance(node, Column) and node.column_name.startswith(
+                self.__nested_col
+            ):
+                return True
+        return False
+
     def process_query(self, query: Query, request_settings: RequestSettings) -> None:
         conditions = query.get_conditions()
         if not conditions:
@@ -135,6 +144,29 @@ class NestedFieldConditionOptimizer(QueryProcessor):
                         return
             if not apply_optimization:
                 return
+
+        # Do not use flattened tags if tags are being unpacked anyway. In that case
+        # using flattened tags only implies loading an additional column thus making
+        # the query heavier and slower
+        if query.get_arrayjoin_from_ast() and self.__find_tags(
+            query.get_arrayjoin_from_ast()
+        ):
+            return
+        if query.get_groupby_from_ast():
+            for expression in query.get_groupby_from_ast():
+                if self.__find_tags(expression):
+                    return
+        if query.get_having_from_ast() and self.__find_tags(
+            query.get_having_from_ast()
+        ):
+            return
+        if not query.get_groupby_from_ast():
+            # If we are not grouping by, the order by can be impactful for this use case
+            # as well. If we have a group by, the order by may not be unpacking all tags.
+            if query.get_orderby_from_ast():
+                for orderby in query.get_orderby_from_ast():
+                    if self.__find_tags(orderby.node):
+                        return
 
         new_conditions = []
         positive_like_expression: List[str] = []