This is the perf of a massive query with tags filtering and group by a tag
> 51 rows in set. Elapsed: 110.565 sec. Processed 126.04 million rows, 217.14 GB (1.14 million rows/s., 1.96 GB/s.)

Same query without flattened tags

> 51 rows in set. Elapsed: 81.081 sec. Processed 126.08 million rows, 133.23 GB (1.55 million rows/s., 1.64 GB/s.)

If we are already unpacking the tags (for the group by), loading an additional big column (the flattened tags) makes things worse since that column is not needed otherwise for the query.